HuggingFace/Transformers-Agent/requirements.txt
huggingface-hub==0.14.1
transformers==4.29.1
Pillow==9.5.0
sounddevice==0.4.6
pdfplumber==0.9.0